# servers/node/src/docs/openapi.yaml
openapi: "3.0.0"
info:
  version: 12.0.0
  title: Jeera Project Manager
  license:
    name: MIT
  contact:
    name: Antoine Cheron
    email: cheron.antoine@gmail.com
paths:
...
  /task/{taskId}:
    parameters: 
      - $ref: '#/components/parameters/taskId'
    get:
      summary: Get the detail of a task
      operationId: getTaskDetails
      tags:
        - tasks
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          links:
            update:
              operationId: updateTask
              parameters:
                taskId: '$response.body#/id'
                parentProjectId: '$response.body#/parentProjectId'
                id: '$response.body#/id'
                title: '$response.body#/title'
                description: '$response.body#/details/description'
                assignee: '$response.body#/details/assignee'
                points: '$response.body#/points'
                tags: '$response.body#/tags'
                priority: '$response.body#/priority'
            reverseArchivedState:
              operationId: reverseTaskArchivedState
              parameters:
                taskId: '$response.body#/id'
            delete:
              operationId: deleteTask
              parameters:
                taskId: '$response.body#/id'
            updateStatus:
              operationId: updateTaskStatus
              parameters:
                taskId: '$response.body#/id'
            analytics:
              operationId: getAnalytics
              parameters:
                resourceId: '$response.body#/id'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a task
      operationId: deleteTask
      tags:
        - tasks
      responses:
        '204':
          description: Success
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /task:
    put:
      summary: Update some fields of the task
      operationId: updateTask
      tags:
        - tasks
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/TechnicalStoryUpdateRequest'
                - $ref: '#/components/schemas/UserStoryUpdateRequest'
      responses:
        '200':
          description: Update successfully done
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          links:
            update:
              operationId: updateTask
              parameters:
                taskId: '$response.body#/id'
                parentProjectId: '$response.body#/parentProjectId'
                id: '$response.body#/id'
                title: '$response.body#/title'
                description: '$response.body#/details/description'
                assignee: '$response.body#/details/assignee'
                points: '$response.body#/points'
                tags: '$response.body#/tags'
                priority: '$response.body#/priority'
            reverseArchivedState:
              operationId: reverseTaskArchivedState
              parameters:
                taskId: '$response.body#/id'
            delete:
              operationId: deleteTask
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
            updateStatus:
              operationId: updateTaskStatus
              parameters:
                taskId: '$response.body#/id'
            analytics:
              operationId: getAnalytics
              parameters:
                resourceId: '$response.body#/id'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /task/{taskId}/archive:
    post:
      description: Reverse the archived status of a task
      operationId: reverseTaskArchivedState
      tags:
        - tasks
      parameters: 
        - name: taskId
          in: path
          description: Id of the task to archive
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  isArchived:
                    type: boolean
                required: [ isArchived ]
          links:
            update:
              operationId: updateTask
              parameters:
                taskId: '$response.body#/id'
                id: '$response.body#/id'
                title: '$response.body#/title'
                description: '$response.body#/details/description'
                assignee: '$response.body#/details/assignee'
                points: '$response.body#/points'
                tags: '$response.body#/details/tags'
                priority: '$response.body#/details/priority'
            reverseArchivedState:
              operationId: reverseTaskArchivedState
              parameters:
                taskId: '$response.body#/id'
            delete:
              operationId: deleteTask
              parameters:
                taskId: '$response.body#/id'
            updateStatus:
              operationId: updateTaskStatus
              parameters:
                taskId: '$response.body#/id'
            analytics:
              operationId: getAnalytics
              x-@relation: vnd_jeera:rel/analytics
              parameters:
                resourceId: '$response.body#/id'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
...
  /tasks:
    parameters: 
        - $ref: '#/components/parameters/projectId'
    get:
      summary: List user tasks
      operationId: listTasks
      tags:
        - tasks
      parameters:
        - $ref: '#/components/parameters/queryProjectId'
        - name: createdBefore
          in: query
          description: Minimum date of tasks to list
          required: false
          schema:
            type: string
            format: date
        - name: offset
          in: query
          description: The number of items to skip before starting to collect the result set
          required: false
          schema:
            type: integer
            format: int8
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Number of elements to get
          required: false
          schema:
            type: integer
            format: int8
            minimum: 1
            maximum: 40
            default: 10
      responses:
        '200':
          description: Success - list of tasks
          headers:
            Link:
              description: 'Contextually relevant links - provided links all have a rel attribute as recommended here: https://www.w3.org/wiki/LinkHeader'
              schema:
                type: string
                format: link-http-header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TasksList'
          links:
            createTechnicalStory:
              operationId: createTechnicalStory
              parameters:
                parentProjectId: "$request.query.queryProjectId"
            createUserStory:
              operationId: createUserStory
              parameters:
                parentProjectId: "$request.query.queryProjectId"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /tasks/technicalStory:
    post:
      summary: Creates a technical story
      operationId: createTechnicalStory
      tags:
        - tasks
      requestBody:
        description: Technical story to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TechnicalStoryCreationRequest'
      responses:
        '201':
          description: Created technical story
          headers:
            Location:
              description: 'URL to the resource created'
              schema:
                type: string
                format: url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicalStory'
          links:
            update:
              operationId: updateTask
              parameters:
                taskId: '$response.body#/id'
                parentProjectId: '$response.body#/parentProjectId'
                id: '$response.body#/id'
                title: '$response.body#/title'
                description: '$response.body#/details/description'
                assignee: '$response.body#/details/assignee'
                points: '$response.body#/points'
                tags: '$response.body#/tags'
                priority: '$response.body#/priority'
            reverseArchivedState:
              operationId: reverseTaskArchivedState
              parameters:
                taskId: '$response.body#/id'
            delete:
              operationId: deleteTask
              parameters:
                taskId: '$response.body#/id'
            updateStatus:
              operationId: updateTaskStatus
              parameters:
                taskId: '$response.body#/id'
            analytics:
              operationId: getAnalytics
              parameters:
                resourceId: '$response.body#/id'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /tasks/userStory:
    parameters: 
      - $ref: '#/components/parameters/projectId'
    post:
      summary: Creates a user story
      operationId: createUserStory
      tags:
        - tasks
      requestBody:
        description: User story to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserStoryCreationRequest'
      responses:
        '201':
          description: Created user story
          headers:
            Location:
              description: 'URL to the resource created'
              schema:
                type: string
                format: url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStory'
          links:
            update:
              operationId: updateTask
              parameters:
                taskId: '$response.body#/id'
                parentProjectId: '$response.body#/parentProjectId'
                id: '$response.body#/id'
                title: '$response.body#/title'
                description: '$response.body#/details/description'
                assignee: '$response.body#/details/assignee'
                points: '$response.body#/points'
                tags: '$response.body#/tags'
                priority: '$response.body#/priority'
            reverseArchivedState:
              operationId: reverseTaskArchivedState
              parameters:
                taskId: '$response.body#/id'
            delete:
              operationId: deleteTask
              parameters:
                taskId: '$response.body#/id'
            updateStatus:
              operationId: updateTaskStatus
              parameters:
                taskId: '$response.body#/id'
            analytics:
              operationId: getAnalytics
              parameters:
                resourceId: '$response.body#/id'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /task/{taskId}/status:
    parameters:
      - $ref: '#/components/parameters/taskId'
    put:
      description: Update the task's status
      operationId: updateTaskStatus
      tags:
        - tasks
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
      responses:
        '204':
          description: Success
          links:
            modify:
              operationId: updateTask
              parameters:
                taskId: '$response.body#/id'
                parentProjectId: '$response.body#/parentProjectId'
                id: '$response.body#/id'
                title: '$response.body#/title'
                description: '$response.body#/details/description'
                assignee: '$response.body#/details/assignee'
                points: '$response.body#/points'
                tags: '$response.body#/tags'
                priority: '$response.body#/priority'
            reverseArchivedState:
              operationId: reverseTaskArchivedState
              parameters:
                taskId: '$response.body#/id'
            updateStatus:
              operationId: updateTaskStatus
              parameters:
                taskId: '$response.body#/id'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
...
    taskId:
      name: taskId
      in: path
      description: Id of the task to retrieve details for
      required: true
      schema:
        type: string
        format: url
    projectId:
      name: projectId
      in: path
      description: Id of the project to retrieve details for
      required: true
      schema:
        type: string
        format: url
...
    queryProjectId:
      name: queryProjectId
      in: query
      description: Id of the project to retrieve details for
      required: true
      schema:
        type: string
        format: url
  schemas:
...
    TasksList:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
    TechnicalStoryCreationRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 4
          maxLength: 80
        parentProjectId:
          type: string
          format: url
        description:
          type: string
          minLength: 0
          maxLength: 4000
        assignee:
          type: string
        tags:
          type: array
          items:
            type: string
          maxItems: 10
        priority:
          $ref: '#/components/schemas/Priority'
      required: [ title, assignee ]
    TechnicalStoryUpdateRequest:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
          minLength: 4
          maxLength: 80
        details:
          $ref: '#/components/schemas/TaskDetailsUpdate'
        tags:
          type: array
          items:
            type: string
          maxItems: 10
        priority:
          $ref: '#/components/schemas/Priority'
    TechnicalStory:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        parentProjectId:
          type: string
        details:
          $ref: '#/components/schemas/TaskDetails'
        tags:
          type: array
          items:
            type: string
        priority:
          $ref: '#/components/schemas/Priority'
        isArchived:
          type: boolean       
      required: [ id, title, parentProjectId, details, isArchived ]
    TaskDetails:
      type: object
      properties:
        description:
          type: string
        assignee:
          type: string
        status:
          type: string
    TaskDetailsUpdate:
      type: object
      properties:
        description:
          type: string
        assignee:
          type: string
    UserStoryCreationRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 4
          maxLength: 80
        parentProjectId:
          type: string
          format: url
        description:
          type: string
          minLength: 0
          maxLength: 4000
        assignee:
          type: string
        points:
          type: number
          minimum: 0
          maximum: 120
        tags:
          type: array
          items:
            type: string
          maxItems: 10
        priority:
          $ref: '#/components/schemas/Priority'
      required: [ title, assignee, points ]
    UserStoryUpdateRequest:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
          minLength: 4
          maxLength: 80
        details:
          $ref: '#/components/schemas/TaskDetailsUpdate'
        points:
          type: number
          minimum: 0
          maximum: 120
        tags:
          type: array
          items:
            type: string
          maxItems: 10
        priority:
          $ref: '#/components/schemas/Priority'
    UserStory:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        points:
          type: number
        parentProjectId:
          type: string
        details:
          $ref: '#/components/schemas/TaskDetails'
        tags:
          type: array
          items:
            type: string
        priority:
          $ref: '#/components/schemas/Priority'
        isArchived:
          type: boolean
      required: [ id, title, points, parentProjectId, details, isArchived ]
    Task:
      oneOf:
        - $ref: '#/components/schemas/TechnicalStory'
        - $ref: '#/components/schemas/UserStory'
...
    Priority:
      type: string
      enum:
        - blocking
        - critical
        - important
        - high
        - medium
        - low
        - simple
...
    Error:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
...